code cleanup for tokenomics report
